Separated out quest sidebar component
